+++ genericsuite/lib/services/crud_editor_sf_filters.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
 
-// import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-
 import 'crud_editor_commons.dart';
 import 'utilities.dart';
 
@@ -12,7 +10,6 @@ Future<Map<String, dynamic>> mandatoryFiltersDbListPreRead(
   Map<String, dynamic> editorConfig,
   String action,
   Map<String, dynamic> params,
-  // FlutterSecureStorage storage,
   BuildContext? context,
 ) async {
   Map<String, dynamic> result = genericFuncArrayDefaultValue(data);
@@ -36,7 +33,6 @@ Future<Map<String, dynamic>> mandatoryFiltersDbPreRead(
   Map<String, dynamic> editorConfig,
   String action,
   Map<String, dynamic> params,
-  // FlutterSecureStorage storage,
   BuildContext? context,
 ) async {
   Map<String, dynamic> result = genericFuncArrayDefaultValue(data);