+++ genericsuite/lib/services/current_user_service.dart
@@ -29,11 +29,9 @@ String getBestApiKey(List<dynamic> userDataApiKeys) {
   return result;
 }
 
-// Future<Map<String, dynamic>> getCurrentUserData(FlutterSecureStorage storage) {
 Future<Map<String, dynamic>> getCurrentUserData() {
   FlutterSecureStorage storage = locator<FlutterSecureStorage>();
 
-  // return loadConfig(storage).then((configStr) {
   return loadConfig().then((configStr) {
     Map<String, dynamic> config = json.decode(configStr);
     Map<String, dynamic> userData = {
@@ -59,14 +57,12 @@ Future<Map<String, dynamic>> getCurrentUserData() {
     if (cusDebug) {
       logDebug('getCurrentUserData | apiUrlUsersGetData: $apiUrlUsersGetData');
     }
-    // var api = HttpUtilities(storage);
     var api = HttpUtilities();
     return api.httpsCall("get", apiUrlUsersGetData, {}, {}, {}).then((data) {
       if (data['error']) {
         return logError(
           'getCurrentUserData | ERROR [1] | data: ${data.toString()}',
           'GCUD-E010',
-          // storage,
         ).then((_) {
           userData['error'] = true;
           userData['errorMessage'] = data['error_message'];
@@ -105,7 +101,6 @@ Future<Map<String, dynamic>> getCurrentUserData() {
               logError(
                 'getCurrentUserData | ERROR [2] | data: ${data.toString()}',
                 'GCUD-E020',
-                // storage,
               ).then((_) {
                 // Pass
               });
@@ -141,7 +136,6 @@ Future<Map<String, dynamic>> getCurrentUserData() {
             logError(
               'getCurrentUserData | ERROR [3] | error: $error',
               'GCUD-E030',
-              // storage,
             ).then((_) {
               // Pass
             });