+++ genericsuite/lib/services/form_field_service.dart
@@ -3,8 +3,6 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-// import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-
 import 'autocomplete_service.dart';
 import 'select_options_service.dart';
 import 'utilities.dart';
@@ -158,7 +156,6 @@ class _PasswordFieldState extends State<PasswordField> {
 }
 
 class DataFormBody extends StatefulWidget {
-  // final FlutterSecureStorage storage;
   final Map<String, dynamic> editorConfig;
   final Map<String, dynamic> constants;
   final Map<String, dynamic> selectedItem;
@@ -172,7 +169,6 @@ class DataFormBody extends StatefulWidget {
 
   const DataFormBody({
     Key? key,
-    // required this.storage,
     required this.editorConfig,
     required this.constants,
     required this.selectedItem,
@@ -574,7 +570,6 @@ class _DataFormBodyState extends State<DataFormBody> {
         case 'suggestion_dropdown':
           formFields.add(
             SuggestionDropdown(
-              // storage: widget.storage,
               config: fieldElement,
               value: fieldElementValue,
               onSelected: (value) {