+++ genericsuite/lib/services/http_service.dart
@@ -15,9 +15,6 @@ const debugConfigValues = false;
 const useServerLog = true;
 
 class HttpUtilities {
-  // final FlutterSecureStorage storage;
-
-  // HttpUtilities(this.storage);
   HttpUtilities();
 
   String jwtToken = '';
@@ -31,7 +28,6 @@ class HttpUtilities {
     Map<String, dynamic> getParams, {
     bool useServerLog = true,
   }) async {
-    // return loadConfig(storage).then((configStr) async {
     return loadConfig().then((configStr) async {
       Map<String, dynamic> config = json.decode(configStr);
       String logErrorMessage = '';
@@ -141,7 +137,6 @@ class HttpUtilities {
             logErrorMessage =
                 'HttpUtilities | httpsCall | Invalid request type: $requestType';
             if (useServerLog) {
-              // await logError(logErrorMessage, 'HC-E010', storage);
               await logError(logErrorMessage, 'HC-E010');
             } else {
               logErrorRaw(logErrorMessage);
@@ -187,7 +182,6 @@ class HttpUtilities {
         logErrorMessage =
             'HttpUtilities | httpsCall | Error [2]: $errorMessage | response: ${response.toString()}';
         if (useServerLog) {
-          // await logError(logErrorMessage, 'HC-E020', storage);
           await logError(logErrorMessage, 'HC-E020');
         } else {
           logErrorRaw(logErrorMessage);
@@ -250,7 +244,6 @@ Map<String, dynamic> getJwtPayload(String jwtTokenRaw) {
   return jwtPayload;
 }
 
-// Future<String> loadConfig(FlutterSecureStorage storage) async {
 Future<String> loadConfig() async {
   final storage = locator<FlutterSecureStorage>();
   String jwtToken = '';
@@ -295,13 +288,7 @@ Future<String> loadConfig() async {
 
 // Log functions
 
-Future<void> logServer(
-  String message,
-  String errorCode,
-  String logType,
-  // FlutterSecureStorage storage,
-) async {
-  // HttpUtilities api = HttpUtilities(storage);
+Future<void> logServer(String message, String errorCode, String logType) async {
   HttpUtilities api = HttpUtilities();
   String urlSuffix = "logs";
   String requestMethod = "POST";
@@ -329,32 +316,23 @@ Future<void> logServer(
   }
 }
 
-// Future<void> logInfo(String message, FlutterSecureStorage storage) async {
 Future<void> logInfo(String message) async {
   logInfoRaw(message);
   if (useServerLog) {
-    // logServer(message, '', 'INFO', storage);
     logServer(message, '', 'INFO');
   }
 }
 
-// Future<void> logWarning(String message, FlutterSecureStorage storage) async {
 Future<void> logWarning(String message) async {
   logWarningRaw(message);
   if (useServerLog) {
-    // logServer(message, '', 'WARNING', storage);
     logServer(message, '', 'WARNING');
   }
 }
 
-Future<void> logError(
-  String message,
-  String errorCode,
-  // FlutterSecureStorage storage,
-) async {
+Future<void> logError(String message, String errorCode) async {
   logErrorRaw('$message [$errorCode]');
   if (useServerLog) {
-    // logServer(message, errorCode, 'ERROR', storage);
     logServer(message, errorCode, 'ERROR');
   }
 }