+++ genericsuite/lib/services/redirect_service.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
 
-// import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-
 import '../views/homepage.dart';
 import 'app_callables_super.dart';
 import 'utilities.dart';
@@ -10,7 +8,6 @@ const redirectToCrud = false;
 const rdSvDebug = false;
 
 Widget getMainScreen(
-  // storage,
   HomePageBodyBuilder homePageBodyBuilder,
   AlternateWidgetBuilder alternateWidgetBuilder,
   AppCallablesSuper appCallables,
@@ -19,23 +16,16 @@ Widget getMainScreen(
     if (rdSvDebug) {
       logDebug('>>>> Redirect to CRUD...');
     }
-    // return alternateWidgetBuilder(storage);
     return alternateWidgetBuilder();
   } else {
     if (rdSvDebug) {
       logDebug('>>>> Redirect to HomePage...');
     }
-    return HomePage(
-      // storage,
-      homePageBodyBuilder,
-      alternateWidgetBuilder,
-      appCallables,
-    );
+    return HomePage(homePageBodyBuilder, alternateWidgetBuilder, appCallables);
   }
 }
 
 void redirectMainScreen(
-  // FlutterSecureStorage storage,
   BuildContext context,
   HomePageBodyBuilder homePageBodyBuilder,
   AlternateWidgetBuilder alternateBodyBuilder,
@@ -45,7 +35,6 @@ void redirectMainScreen(
     context,
     MaterialPageRoute(
       builder: (context) => getMainScreen(
-        // storage,
         homePageBodyBuilder,
         alternateBodyBuilder,
         appCallables,